+++ cpoint.cpp
@@ -17,7 +17,7 @@ std::ostream& operator<< (std::ostream &out, cPoint a)
 	return out;
 }
 
-cPoint::cPoint (int _x, int _y)
+cPoint::cPoint (double _x, double _y)
 {
 	x = _x;
 	y = _y;
@@ -25,22 +25,22 @@ cPoint::cPoint (int _x, int _y)
 
 int cPoint::point_position (cPoint a, cPoint b)
 {
-	float t = (b.x - a.x) * (y - a.y) - (b.y - a.y) * (x - a.x);
+	double t = (b.x - a.x) * (y - a.y) - (b.y - a.y) * (x - a.x);
 	return (fabs(t) < 1e-6) ? (0) : ((t > 0) ? (1) : (-1)); // -1 - справа, 0 - на его прямой и 1 - слева
 }
 
-float cPoint::length_to (cPoint a)
+double cPoint::length_to (cPoint a)
 {
 	return sqrt(pow(a.x - x, 2) + pow(a.y - y, 2));
 }
 
-float cPoint::length_to (cPoint a, cPoint b)
+double cPoint::length_to (cPoint a, cPoint b)
 {
-	float t;
+	double t;
 	
 	if (((x - a.x) * (b.x - a.x) + (y - a.y) * (b.y - a.y)) * ((x - b.x) * (b.x - a.x) + (y - b.y) * (b.y - a.y)) > 0)
 	{
-		float w = pow(x - b.x, 2) + pow(y - b.y, 2);
+		double w = pow(x - b.x, 2) + pow(y - b.y, 2);
 		t = pow(x - a.x, 2) + pow(y - a.y, 2);
 		
 		if (w < t)
